requirements.txt
streamlit>=1.30.0
google-generativeai>=0.8.0
gspread>=6.0.0
google-auth>=2.25.0
Pillow>=10.0.0
